requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.13.1
djangorestframework-mvt==0.2.5
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
